requirements.txt
# Зависимости для Yandex Cloud
# Тяжелые зависимости (opencv-python, numpy, pillow) исключены для уменьшения размера

# FastAPI и сервер
fastapi==0.109.2
uvicorn[standard]==0.27.0
mangum>=0.17.0
python-multipart==0.0.9

# База данных
sqlalchemy==2.0.27
asyncpg==0.29.0  # Для PostgreSQL
greenlet>=3.0.0  # Требуется для SQLAlchemy async операций

# HTTP клиенты
aiohttp==3.9.3
requests==2.31.0

# Валидация и настройки
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Шаблоны
jinja2==3.1.3

# API клиенты
openai>=1.0.0

# QR код генерация
qrcode[pil]>=7.4.2
pillow>=10.0.0
numpy>=1.24.0

# Исключено для уменьшения размера:
# - opencv-python (не используется в API, только в боте)
# - aiosqlite (SQLite не работает на serverless)
# Примечание: numpy используется для удаления фона из изображений
# - aiogram (бот работает отдельно)
# Примечание: pillow используется для генерации QR-кодов
